organizing java file into package, and making a trashcan feature in the shoppingcart
