+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         android:theme="@style/Theme.Pawfect_Purrchase"
         tools:targetApi="31">
 
-        <activity android:name=".MainActivity"
+        <activity android:name=".activities.MainActivity"
             android:exported="true">
 <!--            <intent-filter>-->
 <!--                <action android:name="android.intent.action.MAIN" />-->
@@ -28,13 +28,13 @@
 <!--            </intent-filter>-->
 
         </activity>
-        <activity android:name=".RegisterActivity" />
-        <activity android:name=".ProfileActivity" />
-        <activity android:name=".AboutUsActivity"/>
-        <activity android:name=".ShoppingCartTabLayout"/>
-        <activity android:name=".ProductDetailActivity"/>
+        <activity android:name=".activities.RegisterActivity" />
+        <activity android:name=".activities.ProfileActivity" />
+        <activity android:name=".activities.AboutUsActivity"/>
+        <activity android:name=".activities.ShoppingCartTabLayout"/>
+        <activity android:name=".activities.ProductDetailActivity"/>
         <activity
-            android:name=".LoginActivity"
+            android:name=".activities.LoginActivity"
 
             android:exported="true">
             <intent-filter>
@@ -44,9 +44,9 @@
             </intent-filter>
 
         </activity>
-        <activity android:name=".EquipmentList" />
-        <activity android:name=".FoodList" />
-        <activity android:name=".ToysList" />
+        <activity android:name=".activities.EquipmentList" />
+        <activity android:name=".activities.FoodList" />
+        <activity android:name=".activities.ToysList" />
 
         <meta-data
             android:name="com.google.android.geo.API_KEY"
