+++ app/src/main/java/com/example/pawfect_purrchase/activities/MainActivity.java
@@ -1,12 +1,9 @@
-package com.example.pawfect_purrchase;
+package com.example.pawfect_purrchase.activities;
 
-import android.annotation.SuppressLint;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.Button;
 import android.widget.ImageView;
-import android.widget.ListView;
 
 import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
@@ -14,12 +11,9 @@ import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
 
-import java.util.ArrayList;
-import java.util.List;
+import com.example.pawfect_purrchase.R;
 
 public class MainActivity extends AppCompatActivity {
-    Button btnMap;
-
     ImageView equipmentBtn, foodBtn, toysBtn;
     ImageView homeInfo;
     ImageView userProfile;
