+++ app/src/main/java/com/example/pawfect_purrchase/activities/RegisterActivity.java
@@ -1,7 +1,6 @@
-package com.example.pawfect_purrchase;
+package com.example.pawfect_purrchase.activities;
 
 import android.os.Bundle;
-import android.os.PersistableBundle;
 import android.text.InputType;
 import android.view.View;
 import android.widget.Button;
@@ -11,6 +10,9 @@ import android.widget.Toast;
 
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.pawfect_purrchase.utils.DatabaseHelper;
+import com.example.pawfect_purrchase.R;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
