+++ app/src/main/java/com/example/pawfect_purrchase/fragments/fragment_cart.java
@@ -1,6 +1,5 @@
-package com.example.pawfect_purrchase;
+package com.example.pawfect_purrchase.fragments;
 
-import android.annotation.SuppressLint;
 import android.database.Cursor;
 import android.os.Bundle;
 
@@ -9,11 +8,15 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ArrayAdapter;
 import android.widget.ImageView;
 import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import com.example.pawfect_purrchase.R;
+import com.example.pawfect_purrchase.models.ProductModel;
+import com.example.pawfect_purrchase.utils.CartItemAdapter;
+import com.example.pawfect_purrchase.utils.DatabaseHelper;
 
 import java.util.ArrayList;
 import java.util.List;
