+++ app/src/main/java/com/example/pawfect_purrchase/fragments/fragment_history.java
@@ -1,4 +1,4 @@
-package com.example.pawfect_purrchase;
+package com.example.pawfect_purrchase.fragments;
 
 import android.database.Cursor;
 import android.os.Bundle;
@@ -10,6 +10,11 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.ListView;
+
+import com.example.pawfect_purrchase.R;
+import com.example.pawfect_purrchase.models.TransactionModel;
+import com.example.pawfect_purrchase.utils.DatabaseHelper;
+import com.example.pawfect_purrchase.utils.HistoryAdapter;
 
 import java.util.ArrayList;
 import java.util.List;
