+++ app/src/main/java/com/example/pawfect_purrchase/models/TransactionModel.java
@@ -1,4 +1,4 @@
-package com.example.pawfect_purrchase;
+package com.example.pawfect_purrchase.models;
 
 public class TransactionModel {
     private String transactionCount;
