+++ app/src/main/java/com/example/pawfect_purrchase/utils/ContentAdapter.java
@@ -1,4 +1,4 @@
-package com.example.pawfect_purrchase;
+package com.example.pawfect_purrchase.utils;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -7,6 +7,9 @@ import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import com.example.pawfect_purrchase.models.ProductModel;
+import com.example.pawfect_purrchase.R;
 
 import java.util.List;
