+++ app/src/main/java/com/example/pawfect_purrchase/utils/HistoryAdapter.java
@@ -1,15 +1,17 @@
-package com.example.pawfect_purrchase;
+package com.example.pawfect_purrchase.utils;
 
 import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
-import android.widget.ImageView;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+
+import com.example.pawfect_purrchase.R;
+import com.example.pawfect_purrchase.models.TransactionModel;
 
 import java.util.List;
