+++ app/src/main/res/layout/fragment_cart.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragment_cart"
+    tools:context=".fragments.fragment_cart"
     android:background="#FFEFD9">
 
     <androidx.constraintlayout.widget.ConstraintLayout
