+++ app/src/main/res/layout/fragment_history.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".fragment_history"
+    tools:context=".fragments.fragment_history"
     android:background="#FFEFD9">
 
     <ListView
